move telemetry reload code under the telemetry plugin
